TOT-34 : [REFACTOR] ErrorMessage 삭제 후 FailMessage 반영
- 기존 ErrorMessage 삭제 후 병합된 FailMessage 사용.
