+++ backend/src/main/java/com/triportreat/backend/region/error/exception/RecommendedPlacesNotFoundException.java
@@ -1,6 +1,6 @@
 package com.triportreat.backend.region.error.exception;
 
-import static com.triportreat.backend.common.response.ErrorMessage.RECOMMENDED_PLACE_EMPTY;
+import static com.triportreat.backend.common.response.FailMessage.RECOMMENDED_PLACE_EMPTY;
 
 import com.triportreat.backend.common.AbstractException;
 import org.springframework.http.HttpStatus;
