+++ backend/src/main/java/com/triportreat/backend/region/error/exception/RegionNotFoundException.java
@@ -1,6 +1,6 @@
 package com.triportreat.backend.region.error.exception;
 
-import static com.triportreat.backend.common.response.ErrorMessage.REGION_NOT_FOUND;
+import static com.triportreat.backend.common.response.FailMessage.REGION_NOT_FOUND;
 
 import com.triportreat.backend.common.AbstractException;
 import org.springframework.http.HttpStatus;
